+++ index.html
@@ -233,6 +233,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
   <script type="module" src="js/main.js"></script>
-  <script type="module" src="js/popupContoller.js"></script>
+  <script type="module" src="js/popup-controller.js"></script>
 </body>
 </html>
