+++ js/utils.js
@@ -6,13 +6,6 @@ const getRandomNumber = (min, max) => {
   return Math.floor(result);
 };
 
-const setAttributesId = (selector) => {
-  const element = document.querySelectorAll(selector);
-  for (let i = 0; i < element.length; i++) {
-    element[i].dataset.id = i + 1;
-  }
-};
-
 const createIdGenerator = () => {
   let lastGeneratedId = 0;
 
@@ -28,4 +21,4 @@ const descriptionId = createIdGenerator();
 const photoId = createIdGenerator();
 const commentId = createIdGenerator();
 
-export { getRandomNumber, descriptionId, photoId, commentId, isEscapeKey, setAttributesId};
+export { getRandomNumber, descriptionId, photoId, commentId, isEscapeKey };
